Move eventdetails into its own package; move tabs to their own files
