+++ app/src/main/kotlin/com/thebluealliance/android/ui/events/detail/tabs/EventAlliancesTab.kt
@@ -1,0 +1,50 @@
+package com.thebluealliance.android.ui.events.detail.tabs
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.thebluealliance.android.domain.model.Alliance
+import com.thebluealliance.android.ui.common.EmptyBox
+import com.thebluealliance.android.ui.common.LoadingBox
+
+@Composable
+fun EventAlliancesTab(alliances: List<Alliance>?) {
+    if (alliances == null) {
+        LoadingBox()
+        return
+    }
+    if (alliances.isEmpty()) {
+        EmptyBox("No alliances")
+        return
+    }
+    LazyColumn(modifier = Modifier.fillMaxSize()) {
+        items(alliances, key = { "${it.eventKey}_${it.number}" }) { alliance ->
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 16.dp, vertical = 8.dp),
+            ) {
+                Text(
+                    text = "Alliance ${alliance.number}",
+                    style = MaterialTheme.typography.titleSmall,
+                    fontWeight = FontWeight.Bold,
+                )
+                Text(
+                    text = alliance.picks.joinToString(", ") { it.removePrefix("frc") },
+                    style = MaterialTheme.typography.bodyMedium,
+                    modifier = Modifier.padding(top = 2.dp),
+                )
+            }
+        }
+    }
+}
+
