+++ app/src/main/kotlin/com/thebluealliance/android/ui/events/detail/tabs/EventMatchesTab.kt
@@ -1,0 +1,11 @@
+package com.thebluealliance.android.ui.events.detail.tabs
+
+import androidx.compose.runtime.Composable
+import com.thebluealliance.android.domain.model.Match
+import com.thebluealliance.android.ui.components.MatchList
+
+@Composable
+fun EventMatchesTab(matches: List<Match>?, onNavigateToMatch: (String) -> Unit) {
+    MatchList(matches = matches, onNavigateToMatch = onNavigateToMatch)
+}
+
